requirements.txt
click
rich
python-dotenv
pandas
polars
numpy
pyarrow
clickhouse-driver
clickhouse-connect
sqlalchemy
sqlalchemy-clickhouse
scikit-learn
xgboost
statsmodels
prophet
pyod
shap
imbalanced-learn
great-expectations
pandera
pydantic
plotly
seaborn
matplotlib
dash
anthropic
mcp  # Если доступен, иначе пропустить
langchain
langchain-anthropic
ipywidgets
tqdm
pytest
black
pylint
loguru
jupyterlab-git
jupyterlab-lsp
jupyter-ai